Move Path and SFEN into types
