+++ src/models/index.ts
@@ -3,5 +3,4 @@ export { JumpTo } from "./jumpTo";
 export type { JumpMap } from "./jumpTo";
 export { KifuTree } from "./kifuTree";
 export { KifuTreeNode } from "./kifuTreeNode";
-export type { Path } from "./path";
-export type { SFEN } from "./sfen";
+export type { Path, SFEN } from "./types";
