+++ src/models/types.ts
@@ -1,3 +1,5 @@
 import { List } from "immutable";
 
 export type Path = List<number>;
+
+export type SFEN = string;
